pyproject.toml
[project]
name = "nasa-mcp-my"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "mcp>=1.17.0",
    "notebook>=7.4.7",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]
